move opa calls out of controller
