+++ src/main/java/tech/stackable/hbase/OpenPolicyAgentAccessController.java
@@ -1,18 +1,8 @@
 package tech.stackable.hbase;
 
-import com.fasterxml.jackson.annotation.JsonAutoDetect;
-import com.fasterxml.jackson.annotation.PropertyAccessor;
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.DeserializationFeature;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.SerializationFeature;
 import com.google.protobuf.RpcCallback;
 import com.google.protobuf.RpcController;
 import java.io.IOException;
-import java.net.URI;
-import java.net.http.HttpClient;
-import java.net.http.HttpRequest;
-import java.net.http.HttpResponse;
 import java.util.*;
 import org.apache.hadoop.hbase.*;
 import org.apache.hadoop.hbase.client.*;
@@ -24,11 +14,9 @@ import org.apache.hadoop.hbase.security.UserProvider;
 import org.apache.hadoop.hbase.security.access.*;
 import org.apache.hadoop.hbase.security.access.Permission.Action;
 import org.apache.hadoop.hbase.wal.WALEdit;
-import org.apache.hadoop.security.AccessControlException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import tech.stackable.hbase.opa.OpaAllowQuery;
-import tech.stackable.hbase.opa.OpaException;
+import tech.stackable.hbase.opa.OpaAclChecker;
 
 public class OpenPolicyAgentAccessController
     implements MasterCoprocessor,
@@ -46,21 +34,10 @@ public class OpenPolicyAgentAccessController
   private boolean authorizationEnabled;
   private boolean cellFeaturesEnabled;
 
+  private OpaAclChecker opaAclChecker;
+
   // Opa-related
   public static final String OPA_POLICY_URL_PROP = "hbase.security.authorization.opa.policy.url";
-  private final HttpClient httpClient = HttpClient.newHttpClient();
-  private URI opaUri;
-  private final ObjectMapper json;
-
-  public OpenPolicyAgentAccessController() {
-    this.json =
-        new ObjectMapper()
-            .configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false)
-            .configure(SerializationFeature.FAIL_ON_EMPTY_BEANS, false)
-            .setVisibility(PropertyAccessor.ALL, JsonAutoDetect.Visibility.NONE)
-            .setVisibility(PropertyAccessor.FIELD, JsonAutoDetect.Visibility.PUBLIC_ONLY)
-            .setVisibility(PropertyAccessor.GETTER, JsonAutoDetect.Visibility.PUBLIC_ONLY);
-  }
 
   @Override
   public void start(CoprocessorEnvironment env) throws IOException {
@@ -83,21 +60,10 @@ public class OpenPolicyAgentAccessController
     this.userProvider = UserProvider.instantiate(env.getConfiguration());
 
     // opa-related
-    if (authorizationEnabled) {
-      String opaPolicyUrl = env.getConfiguration().get(OPA_POLICY_URL_PROP);
-      if (opaPolicyUrl == null) {
-        throw new OpaException.UriMissing(OPA_POLICY_URL_PROP);
-      }
-
-      try {
-        this.opaUri = URI.create(opaPolicyUrl);
-      } catch (Exception e) {
-        throw new OpaException.UriInvalid(opaUri, e);
-      }
-    }
+    this.opaAclChecker =
+        new OpaAclChecker(authorizationEnabled, env.getConfiguration().get(OPA_POLICY_URL_PROP));
   }
 
-  // TODO replace with user that returns the whole name for getShortName()
   private User getActiveUser(ObserverContext<?> ctx) throws IOException {
     // for non-rpc handling, fallback to system user
     Optional<User> optionalUser = ctx.getCaller();
@@ -105,70 +71,6 @@ public class OpenPolicyAgentAccessController
       return optionalUser.get();
     }
     return userProvider.getCurrent();
-  }
-
-  private void checkPermissionInfo(User user, TableName table, Action action)
-      throws AccessControlException {
-    OpaAllowQuery query =
-        new OpaAllowQuery(new OpaAllowQuery.OpaAllowQueryInput(user.getUGI(), table, action));
-
-    String body;
-    try {
-      body = json.writeValueAsString(query);
-    } catch (JsonProcessingException e) {
-      throw new OpaException.SerializeFailed(e);
-    }
-
-    String prettyPrinted;
-    try {
-      prettyPrinted = json.writerWithDefaultPrettyPrinter().writeValueAsString(query);
-    } catch (JsonProcessingException e) {
-      LOG.error(
-          "Could not pretty print the following request body (printing raw version instead): {}",
-          body);
-      throw new OpaException.SerializeFailed(e);
-    }
-
-    LOG.info("Request body:\n{}", prettyPrinted);
-    HttpResponse<String> response = null;
-    try {
-      response =
-          httpClient.send(
-              HttpRequest.newBuilder(opaUri)
-                  .header("Content-Type", "application/json")
-                  .POST(HttpRequest.BodyPublishers.ofString(body))
-                  .build(),
-              HttpResponse.BodyHandlers.ofString());
-      LOG.debug("Opa response: {}", response.body());
-    } catch (Exception e) {
-      LOG.error(e.getMessage());
-      throw new OpaException.QueryFailed(e);
-    }
-
-    switch (Objects.requireNonNull(response).statusCode()) {
-      case 200:
-        break;
-      case 404:
-        throw new OpaException.EndPointNotFound(opaUri.toString());
-      default:
-        throw new OpaException.OpaServerError(query.toString(), response);
-    }
-
-    OpaQueryResult result;
-    try {
-      result = json.readValue(response.body(), OpaQueryResult.class);
-    } catch (JsonProcessingException e) {
-      throw new OpaException.DeserializeFailed(e);
-    }
-
-    if (result.result == null || !result.result) {
-      throw new AccessControlException("OPA denied the request");
-    }
-  }
-
-  private static class OpaQueryResult {
-    // Boxed Boolean to detect not-present vs explicitly false
-    public Boolean result;
   }
 
   @Override
@@ -189,10 +91,8 @@ public class OpenPolicyAgentAccessController
     User user = getActiveUser(c);
     LOG.info("prePut: start with [{}]", user);
 
-    if (authorizationEnabled) {
-      checkPermissionInfo(
-          user, c.getEnvironment().getRegion().getRegionInfo().getTable(), Action.WRITE);
-    }
+    opaAclChecker.checkPermissionInfo(
+        user, c.getEnvironment().getRegion().getRegionInfo().getTable(), Action.WRITE);
   }
 
   @Override
